third_party/blink/web_tests/external/wpt/url/toascii.window-expected.txt

This is a testharness.js-based test.
Found 162 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] xn--a (using URL)
  assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
[FAIL] xn--a (using URL.host)
  assert_equals: expected "x" but got "xn--a"
[FAIL] xn--a (using URL.hostname)
  assert_equals: expected "x" but got "xn--a"
[FAIL] xn--a (using <a>)
  assert_equals: expected "" but got "xn--a"
[FAIL] xn--a (using <a>.host)
  assert_equals: expected "x" but got "xn--a"
[FAIL] xn--a (using <a>.hostname)
  assert_equals: expected "x" but got "xn--a"
[FAIL] xn--a (using <area>)
  assert_equals: expected "" but got "xn--a"
[FAIL] xn--a (using <area>.host)
  assert_equals: expected "x" but got "xn--a"
[FAIL] xn--a (using <area>.hostname)
  assert_equals: expected "x" but got "xn--a"
[FAIL] xn--a.xn--zca (using URL)
  assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
[FAIL] xn--a.xn--zca (using URL.host)
  assert_equals: expected "x" but got "xn--a.xn--zca"
[FAIL] xn--a.xn--zca (using URL.hostname)
  assert_equals: expected "x" but got "xn--a.xn--zca"
[FAIL] xn--a.xn--zca (using <a>)
  assert_equals: expected "" but got "xn--a.xn--zca"
[FAIL] xn--a.xn--zca (using <a>.host)
  assert_equals: expected "x" but got "xn--a.xn--zca"
[FAIL] xn--a.xn--zca (using <a>.hostname)
  assert_equals: expected "x" but got "xn--a.xn--zca"
[FAIL] xn--a.xn--zca (using <area>)
  assert_equals: expected "" but got "xn--a.xn--zca"
[FAIL] xn--a.xn--zca (using <area>.host)
  assert_equals: expected "x" but got "xn--a.xn--zca"
[FAIL] xn--a.xn--zca (using <area>.hostname)
  assert_equals: expected "x" but got "xn--a.xn--zca"
[FAIL] xn--ls8h= (using URL)
  assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
[FAIL] xn--ls8h= (using URL.host)
  assert_equals: expected "x" but got "xn--ls8h="
[FAIL] xn--ls8h= (using URL.hostname)
  assert_equals: expected "x" but got "xn--ls8h="
[FAIL] xn--ls8h= (using <a>)
  assert_equals: expected "" but got "xn--ls8h="
[FAIL] xn--ls8h= (using <a>.host)
  assert_equals: expected "x" but got "xn--ls8h="
[FAIL] xn--ls8h= (using <a>.hostname)
  assert_equals: expected "x" but got "xn--ls8h="
[FAIL] xn--ls8h= (using <area>)
  assert_equals: expected "" but got "xn--ls8h="
[FAIL] xn--ls8h= (using <area>.host)
  assert_equals: expected "x" but got "xn--ls8h="
[FAIL] xn--ls8h= (using <area>.hostname)
  assert_equals: expected "x" but got "xn--ls8h="
[FAIL] ‍.example (using URL)
  assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
[FAIL] ‍.example (using URL.host)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] ‍.example (using URL.hostname)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] ‍.example (using <a>)
  assert_equals: expected "" but got "xn--1ug.example"
[FAIL] ‍.example (using <a>.host)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] ‍.example (using <a>.hostname)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] ‍.example (using <area>)
  assert_equals: expected "" but got "xn--1ug.example"
[FAIL] ‍.example (using <area>.host)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] ‍.example (using <area>.hostname)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] xn--1ug.example (using URL)
  assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
[FAIL] xn--1ug.example (using URL.host)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] xn--1ug.example (using URL.hostname)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] xn--1ug.example (using <a>)
  assert_equals: expected "" but got "xn--1ug.example"
[FAIL] xn--1ug.example (using <a>.host)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] xn--1ug.example (using <a>.hostname)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] xn--1ug.example (using <area>)
  assert_equals: expected "" but got "xn--1ug.example"
[FAIL] xn--1ug.example (using <area>.host)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] xn--1ug.example (using <area>.hostname)
  assert_equals: expected "x" but got "xn--1ug.example"
[FAIL] xn--a-yoc (using URL)
  assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
[FAIL] xn--a-yoc (using URL.host)
  assert_equals: expected "x" but got "xn--a-yoc"
[FAIL] xn--a-yoc (using URL.hostname)
  assert_equals: expected "x" but got "xn--a-yoc"
[FAIL] xn--a-yoc (using <a>)
  assert_equals: expected "" but got "xn--a-yoc"
[FAIL] xn--a-yoc (using <a>.host)
  assert_equals: expected "x" but got "xn--a-yoc"
[FAIL] xn--a-yoc (using <a>.hostname)
  assert_equals: expected "x" but got "xn--a-yoc"
[FAIL] xn--a-yoc (using <area>)
  assert_equals: expected "" but got "xn--a-yoc"
[FAIL] xn--a-yoc (using <area>.host)
  assert_equals: expected "x" but got "xn--a-yoc"
[FAIL] xn--a-yoc (using <area>.hostname)
  assert_equals: expected "x" but got "xn--a-yoc"
[FAIL] xn--zn7c.com (using URL)
  assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
[FAIL] xn--zn7c.com (using URL.host)
  assert_equals: expected "x" but got "xn--zn7c.com"
[FAIL] xn--zn7c.com (using URL.hostname)
  assert_equals: expected "x" but got "xn--zn7c.com"
[FAIL] xn--zn7c.com (using <a>)
  assert_equals: expected "" but got "xn--zn7c.com"
[FAIL] xn--zn7c.com (using <a>.host)
  assert_equals: expected "x" but got "xn--zn7c.com"
[FAIL] xn--zn7c.com (using <a>.hostname)
  assert_equals: expected "x" but got "xn--zn7c.com"
[FAIL] xn--zn7c.com (using <area>)
  assert_equals: expected "" but got "xn--zn7c.com"
[FAIL] xn--zn7c.com (using <area>.host)
  assert_equals: expected "x" but got "xn--zn7c.com"
[FAIL] xn--zn7c.com (using <area>.hostname)
  assert_equals: expected "x" but got "xn--zn7c.com"
[FAIL] ≮ (using URL)
  Failed to construct 'URL': Invalid URL
[FAIL] ≮ (using URL.host)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] ≮ (using URL.hostname)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] ≮ (using <a>)
  Failed to construct 'URL': Invalid URL
[FAIL] ≮ (using <a>.host)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] ≮ (using <a>.hostname)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] ≮ (using <area>)
  Failed to construct 'URL': Invalid URL
[FAIL] ≮ (using <area>.host)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] ≮ (using <area>.hostname)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] ≯ (using URL)
  Failed to construct 'URL': Invalid URL
[FAIL] ≯ (using URL.host)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] ≯ (using URL.hostname)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] ≯ (using <a>)
  Failed to construct 'URL': Invalid URL
[FAIL] ≯ (using <a>.host)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] ≯ (using <a>.hostname)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] ≯ (using <area>)
  Failed to construct 'URL': Invalid URL
[FAIL] ≯ (using <area>.host)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] ≯ (using <area>.hostname)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] <­̸ (using URL)
  Failed to construct 'URL': Invalid URL
[FAIL] <­̸ (using URL.host)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] <­̸ (using URL.hostname)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] <­̸ (using <a>)
  Failed to construct 'URL': Invalid URL
[FAIL] <­̸ (using <a>.host)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] <­̸ (using <a>.hostname)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] <­̸ (using <area>)
  Failed to construct 'URL': Invalid URL
[FAIL] <­̸ (using <area>.host)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] <­̸ (using <area>.hostname)
  assert_equals: expected "xn--gdh" but got "x"
[FAIL] >­̸ (using URL)
  Failed to construct 'URL': Invalid URL
[FAIL] >­̸ (using URL.host)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] >­̸ (using URL.hostname)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] >­̸ (using <a>)
  Failed to construct 'URL': Invalid URL
[FAIL] >­̸ (using <a>.host)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] >­̸ (using <a>.hostname)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] >­̸ (using <area>)
  Failed to construct 'URL': Invalid URL
[FAIL] >­̸ (using <area>.host)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] >­̸ (using <area>.hostname)
  assert_equals: expected "xn--hdh" but got "x"
[FAIL] www .lookout.net (using URL)
  assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
[FAIL] www .lookout.net (using URL.host)
  assert_equals: expected "x" but got "www%20.lookout.net"
[FAIL] www .lookout.net (using URL.hostname)
  assert_equals: expected "x" but got "www%20.lookout.net"
[FAIL] www .lookout.net (using <a>)
  assert_equals: expected "" but got "www%20.lookout.net"
[FAIL] www .lookout.net (using <a>.host)
  assert_equals: expected "x" but got "www%20.lookout.net"
[FAIL] www .lookout.net (using <a>.hostname)
  assert_equals: expected "x" but got "www%20.lookout.net"
[FAIL] www .lookout.net (using <area>)
  assert_equals: expected "" but got "www%20.lookout.net"
[FAIL] www .lookout.net (using <area>.host)
  assert_equals: expected "x" but got "www%20.lookout.net"
[FAIL] www .lookout.net (using <area>.hostname)
  assert_equals: expected "x" but got "www%20.lookout.net"
[FAIL] look᠎out.net (using URL)
  Failed to construct 'URL': Invalid URL
[FAIL] look᠎out.net (using URL.host)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] look᠎out.net (using URL.hostname)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] look᠎out.net (using <a>)
  Failed to construct 'URL': Invalid URL
[FAIL] look᠎out.net (using <a>.host)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] look᠎out.net (using <a>.hostname)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] look᠎out.net (using <area>)
  Failed to construct 'URL': Invalid URL
[FAIL] look᠎out.net (using <area>.host)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] look᠎out.net (using <area>.hostname)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] look⁫out.net (using URL)
  Failed to construct 'URL': Invalid URL
[FAIL] look⁫out.net (using URL.host)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] look⁫out.net (using URL.hostname)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] look⁫out.net (using <a>)
  Failed to construct 'URL': Invalid URL
[FAIL] look⁫out.net (using <a>.host)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] look⁫out.net (using <a>.hostname)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] look⁫out.net (using <area>)
  Failed to construct 'URL': Invalid URL
[FAIL] look⁫out.net (using <area>.host)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] look⁫out.net (using <area>.hostname)
  assert_equals: expected "lookout.net" but got "x"
[FAIL] Ӏ.com (using URL)
  Failed to construct 'URL': Invalid URL
[FAIL] Ӏ.com (using URL.host)
  assert_equals: expected "xn--s5a.com" but got "x"
[FAIL] Ӏ.com (using URL.hostname)
  assert_equals: expected "xn--s5a.com" but got "x"
[FAIL] Ӏ.com (using <a>)
  Failed to construct 'URL': Invalid URL
[FAIL] Ӏ.com (using <a>.host)
  assert_equals: expected "xn--s5a.com" but got "x"
[FAIL] Ӏ.com (using <a>.hostname)
  assert_equals: expected "xn--s5a.com" but got "x"
[FAIL] Ӏ.com (using <area>)
  Failed to construct 'URL': Invalid URL
[FAIL] Ӏ.com (using <area>.host)
  assert_equals: expected "xn--s5a.com" but got "x"
[FAIL] Ӏ.com (using <area>.hostname)
  assert_equals: expected "xn--s5a.com" but got "x"
[FAIL] 㛼.com (using URL)
  Failed to construct 'URL': Invalid URL
[FAIL] 㛼.com (using URL.host)
  assert_equals: expected "xn--snl.com" but got "x"
[FAIL] 㛼.com (using URL.hostname)
  assert_equals: expected "xn--snl.com" but got "x"
[FAIL] 㛼.com (using <a>)
  Failed to construct 'URL': Invalid URL
[FAIL] 㛼.com (using <a>.host)
  assert_equals: expected "xn--snl.com" but got "x"
[FAIL] 㛼.com (using <a>.hostname)
  assert_equals: expected "xn--snl.com" but got "x"
[FAIL] 㛼.com (using <area>)
  Failed to construct 'URL': Invalid URL
[FAIL] 㛼.com (using <area>.host)
  assert_equals: expected "xn--snl.com" but got "x"
[FAIL] 㛼.com (using <area>.hostname)
  assert_equals: expected "xn--snl.com" but got "x"
[FAIL] Ↄ.com (using URL)
  Failed to construct 'URL': Invalid URL
[FAIL] Ↄ.com (using URL.host)
  assert_equals: expected "xn--r5g.com" but got "x"
[FAIL] Ↄ.com (using URL.hostname)
  assert_equals: expected "xn--r5g.com" but got "x"
[FAIL] Ↄ.com (using <a>)
  Failed to construct 'URL': Invalid URL
[FAIL] Ↄ.com (using <a>.host)
  assert_equals: expected "xn--r5g.com" but got "x"
[FAIL] Ↄ.com (using <a>.hostname)
  assert_equals: expected "xn--r5g.com" but got "x"
[FAIL] Ↄ.com (using <area>)
  Failed to construct 'URL': Invalid URL
[FAIL] Ↄ.com (using <area>.host)
  assert_equals: expected "xn--r5g.com" but got "x"
[FAIL] Ↄ.com (using <area>.hostname)
  assert_equals: expected "xn--r5g.com" but got "x"
[FAIL] xn--0.com (using URL)
  assert_throws_js: function "() => makeURL("url", hostTest.input)" did not throw
[FAIL] xn--0.com (using URL.host)
  assert_equals: expected "x" but got "xn--0.com"
[FAIL] xn--0.com (using URL.hostname)
  assert_equals: expected "x" but got "xn--0.com"
[FAIL] xn--0.com (using <a>)
  assert_equals: expected "" but got "xn--0.com"
[FAIL] xn--0.com (using <a>.host)
  assert_equals: expected "x" but got "xn--0.com"
[FAIL] xn--0.com (using <a>.hostname)
  assert_equals: expected "x" but got "xn--0.com"
[FAIL] xn--0.com (using <area>)
  assert_equals: expected "" but got "xn--0.com"
[FAIL] xn--0.com (using <area>.host)
  assert_equals: expected "x" but got "xn--0.com"
[FAIL] xn--0.com (using <area>.hostname)
  assert_equals: expected "x" but got "xn--0.com"
Harness: the test ran to completion.
